moved all scripts outside of the UI/Components folder
