+++ Assets/_Application/_Twin/OpenURLInBrowser.cs
@@ -1,5 +1,3 @@
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 
 namespace Netherlands3D.Twin
