+++ Assets/_Application/_Twin/UI/MatchSpriteToSelectionState.cs
@@ -1,11 +1,7 @@
-using System;
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
-using UnityEngine.EventSystems;
 using UnityEngine.UI;
 
-namespace Netherlands3D.Twin
+namespace Netherlands3D.Twin.UI
 {
     public class MatchSpriteToSelectionState : MonoBehaviour
     {
