+++ Assets/_Application/_Twin/UI/MinMaxWidth.cs
@@ -1,8 +1,6 @@
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 
-namespace Netherlands3D.Twin
+namespace Netherlands3D.Twin.UI
 {
     public class MinMaxWidth : MonoBehaviour
     {
