+++ Assets/_Application/_Twin/UI/TransformHandleButtonsPanel.cs
@@ -1,8 +1,6 @@
-using RuntimeHandle;
 using UnityEngine;
 
-
-namespace Netherlands3D.Twin
+namespace Netherlands3D.Twin.UI
 {
     public class TransformHandleButtonsPanel : MonoBehaviour
     {
