+++ Assets/_Functionalities/Wms/Scripts/UI.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: aca10a9a596724b4da0e9266609e12ce
+guid: 0d3a5e6ece9dc44caa10a407c3d9a702
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
