+++ Assets/_Functionalities/Wms/Scripts/UI/LegendClampHeight.cs
